+++ src/urn_old.hpp
@@ -1,7 +1,7 @@
 /*! 
  * \file urn.hpp
  * \author Ulrich Eisenecker
- * \date June 14, 2023
+ * \date March 10, 2023
  *  
  * Header file of the draw generator
  * 
@@ -19,9 +19,13 @@
 #ifndef URN_HPP
 #define URN_HPP
 
+/*!
+ * Include directives for urn.hpp
+ */
+
 #include <vector>
-#include <iostream> // Because of cerr
-#include <exception> 
+#include <iostream>  // Because of cerr
+#include <cstdlib>   // Because of exit()
 
 /*! 
  * \namespace urn
@@ -31,13 +35,12 @@
 
 namespace urn
 {
-   using uint = unsigned int;       /*!< Using decleration to represent the set of natural numbers. */
-   using Draw = std::vector<uint>;  /*!< Using decleration to represent a combination and permutation. */
+   using uint = unsigned int;          /*!< Using decleration to represent the set of natural numbers. */
+   using Draw = std::vector<uint>;     /*!< Using decleration to represent a combination and permutation. */
 
    /*!
     * \class UrnOR – urn where the order is important and which contains repetitions.
     */
-
    class UrnOR
    {
       public:
@@ -48,7 +51,8 @@ namespace urn
           * Initializes m_balls which creates a vector<> with k elements, each with the value 0.
           * The check parameter is not intended to be specified by the user.
           * The constructor body performs a check that is specific for the UrnOR, if the check value is 1.
-          * If the UrnOR exemplar to be created is invalid, an exception is thrown.
+          * If the UrnOR exemplar to be created is invalid, the entire program is terminated by calling the std::exit() function.
+          * 
           * @param[in] n      The number of balls inside the urn.
           * @param[in] k      The size of a draw from the urn.
           * @param[in] check  The check value.
@@ -60,7 +64,7 @@ namespace urn
           * \return m_n the number of balls inside the urn.
           */
          uint n() const;
-         
+
          /*!
           * \brief Getter method which returns m_k.
           * \return m_k the size of a draw from the urn.
@@ -75,7 +79,7 @@ namespace urn
           * @param[in] index Index of the element to be accessed. 
           */
          uint at(uint index) const;
-         
+
          /*!
           * \brief Overload of the [] index operator
           * Returns the element specified by index as value.
@@ -91,7 +95,7 @@ namespace urn
           * \return true/false depends if there is a next draw.
           */
          [[nodiscard]] virtual bool next();
-         
+
          /*!
           * \brief Virtual destructor for UrnOR.
           * Virtual destructor to be a valid base class.
@@ -108,7 +112,6 @@ namespace urn
    /*!
     * \class UrnO – urn where the order is important and does not include repetitions.
     */
-
    class UrnO: public virtual UrnOR
    {
       public:
@@ -120,14 +123,14 @@ namespace urn
           * The check parameter is not intended to be specified by the user.
           * The constructor body performs a check that is specific for the UrnO, if the check value is 2.
           * The constructor body executes a for loop to initialize m_balls correctly (no repetitions).
-          * If the UrnO exemplar to be created is invalid, an exception is thrown.
+          * If the UrnO exemplar to be created is invalid, the entire program is terminated by calling the std::exit() function.
           * 
           * @param[in] n      The number of balls inside the urn.
           * @param[in] k      The size of a draw from the urn.
           * @param[in] check  The check value.
           */
          explicit UrnO(uint n,uint k,uint check = 2);
-         
+
          /*!
           * \brief Method to generate the next draw (order important, without repetition).
           * Calls UrnOR::next() and additionally checks if the previously generated draw contains a duplicate element.
@@ -158,14 +161,14 @@ namespace urn
           * Constructor calls the constructor of UrnOR to create an exampler.
           * The check parameter is not intended to be specified by the user.
           * The constructor body performs a check that is specific for the UrnR, if the check value is 3.
-          * If the UrnR exemplar to be created is invalid, an exception is thrown.
+          * If the UrnR exemplar to be created is invalid, the entire program is terminated by calling the std::exit() function.
           * 
           * @param[in] n      The number of balls inside the urn.
           * @param[in] k      The size of a draw from the urn.
           * @param[in] check  The check value.
           */
          explicit UrnR(uint n,uint k,uint check = 3);
-         
+
          /*!
           * \brief Method to generate the next draw (order not important, with repetition).
           * Calls UrnOR::next() and additionally checks if an element with a lower index value is greater than its neighbor element with index value plus one.
@@ -175,7 +178,7 @@ namespace urn
           * \return true/false depends if there is a next draw.
           */
          [[nodiscard]] bool next() override;
-         
+
       protected:
          /*!
           * \brief Method to check if an element with a lower index value is greater than its neighbor element with index value plus one.
@@ -209,80 +212,8 @@ namespace urn
           * If there is a valid next draw, the member function returns true, otherwise false.
           * 
           * \return true/false depends if there is a next draw.
-          */                            
+          */
          [[nodiscard]] bool next() override;
    };
-
-   template <typename T,bool ORDER = true,bool REPETITION = true>
-   class GenericUrn
-   {
-      public:
-         GenericUrn(uint k,const std::vector<T>& elements)
-            :m_urn { static_cast<uint>(elements.size()),k },
-             m_elements { elements }
-         {}
-         uint n() const
-         {
-            return m_urn.n();
-         }
-         uint k() const
-         {
-            return m_urn.k();
-         }
-         const T& at(uint index) const
-         {
-            return m_elements.at(m_urn.at(index));
-         }
-         const T& operator[](uint index) const
-         {
-            return m_elements[m_urn[index]];
-         }
-         [[nodiscard]] virtual bool next()
-         {
-            return m_urn.next();
-         }
-                 
-      private:
-         template <bool O,bool R>
-         class SelectUrn
-         {
-            public:
-               using UrnType = UrnOR;
-         };
-         template <>
-         class SelectUrn<true,false>
-         {
-            public:
-               using UrnType = UrnO;
-         };
-         template <>
-         class SelectUrn<false,true>
-         {
-            public:
-               using UrnType = UrnR;
-         };
-         template <>
-         class SelectUrn<false,false>
-         {
-            public:
-               using UrnType = Urn;
-         };
-         using UrnType = typename SelectUrn<ORDER,REPETITION>::UrnType;
-         UrnType m_urn;
-         std::vector<T> m_elements;
-   };
-
-   template <class T> // According to https://en.wikipedia.org/wiki/Permutation
-   using Permutation = GenericUrn<T,true,true>;
-
-   template <class T> // According to https://en.wikipedia.org/wiki/Partial_permutation
-   using PartialPermutation = GenericUrn<T,true,false>;
-
-   template <class T> // According to https://en.wikipedia.org/wiki/Combination
-   using MultiCombination = GenericUrn<T,false,true>;
-
-   template <class T> // According to https://en.wikipedia.org/wiki/Combination
-   using Combination = GenericUrn<T,false,false>;
-
 }
 #endif // URN_HPP
